+++ keys/tx/reader_test.go
@@ -6,9 +6,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	crypto "github.com/tendermint/go-crypto"
-	data "github.com/tendermint/go-wire/data"
+	"github.com/tendermint/go-crypto/keys"
 	"github.com/tendermint/go-crypto/keys/cryptostore"
 	"github.com/tendermint/go-crypto/keys/storage/memstorage"
+	data "github.com/tendermint/go-wire/data"
 )
 
 func TestReader(t *testing.T) {
@@ -18,14 +19,15 @@ func TestReader(t *testing.T) {
 	cstore := cryptostore.New(
 		cryptostore.SecretBox,
 		memstorage.New(),
+		keys.MustLoadCodec("english"),
 	)
 	type sigs struct{ name, pass string }
 	u := sigs{"alice", "1234"}
 	u2 := sigs{"bob", "foobar"}
 
-	_, err := cstore.Create(u.name, u.pass, algo)
+	_, _, err := cstore.Create(u.name, u.pass, algo)
 	require.Nil(err, "%+v", err)
-	_, err = cstore.Create(u2.name, u2.pass, algo)
+	_, _, err = cstore.Create(u2.name, u2.pass, algo)
 	require.Nil(err, "%+v", err)
 
 	cases := []struct {